refactor(ui): TE-1270 moved PageHeader render and props out of SubscriptionGroupWizard and into respective parent components
